# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tmalldedede_agentbox_internal_session.Config:
    properties:
      cpu_limit:
        type: number
      memory_limit:
        type: integer
    type: object
  github_com_tmalldedede_agentbox_internal_session.CreateRequest:
    properties:
      agent:
        type: string
      config:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Config'
      env:
        additionalProperties:
          type: string
        type: object
      profile_id:
        type: string
      workspace:
        type: string
    required:
    - agent
    - workspace
    type: object
  github_com_tmalldedede_agentbox_internal_session.ExecEvent:
    properties:
      raw:
        items:
          type: integer
        type: array
      type:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_session.ExecRequest:
    properties:
      allowed_tools:
        description: 允许的工具列表
        items:
          type: string
        type: array
      disallowed_tools:
        description: 禁用的工具列表
        items:
          type: string
        type: array
      include_events:
        description: 是否返回完整事件列表
        type: boolean
      max_turns:
        description: 最大对话轮数 (1-100, 默认 10)
        type: integer
      prompt:
        type: string
      timeout:
        description: 超时秒数 (10-3600, 默认 300)
        type: integer
    required:
    - prompt
    type: object
  github_com_tmalldedede_agentbox_internal_session.ExecResponse:
    properties:
      error:
        type: string
      events:
        description: 完整事件列表 (当 include_events=true)
        items:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.ExecEvent'
        type: array
      execution_id:
        type: string
      exit_code:
        type: integer
      message:
        description: Agent 最终回复 (新增)
        type: string
      output:
        description: 原始输出 (兼容旧版)
        type: string
      usage:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.TokenUsage'
        description: Token 使用统计
    type: object
  github_com_tmalldedede_agentbox_internal_session.Execution:
    properties:
      ended_at:
        type: string
      error:
        type: string
      exit_code:
        type: integer
      id:
        type: string
      output:
        type: string
      prompt:
        type: string
      session_id:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.ExecutionStatus'
    type: object
  github_com_tmalldedede_agentbox_internal_session.ExecutionStatus:
    enum:
    - pending
    - running
    - success
    - failed
    type: string
    x-enum-varnames:
    - ExecutionPending
    - ExecutionRunning
    - ExecutionSuccess
    - ExecutionFailed
  github_com_tmalldedede_agentbox_internal_session.Session:
    properties:
      agent:
        type: string
      config:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Config'
      container_id:
        type: string
      created_at:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      profile_id:
        type: string
      status:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Status'
      updated_at:
        type: string
      workspace:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_session.Status:
    enum:
    - creating
    - running
    - stopped
    - error
    type: string
    x-enum-varnames:
    - StatusCreating
    - StatusRunning
    - StatusStopped
    - StatusError
  github_com_tmalldedede_agentbox_internal_session.TokenUsage:
    properties:
      cached_input_tokens:
        type: integer
      input_tokens:
        type: integer
      output_tokens:
        type: integer
    type: object
...
  internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      type:
        description: 错误类型
        type: string
    type: object
...
host: localhost:18080
info:
  contact:
    email: support@agentbox.dev
    name: AgentBox Team
    url: https://github.com/user/agentbox
  description: |-
    AgentBox - Open-source AI Agent containerized runtime platform

    ## Overview
    AgentBox provides a unified API for managing AI coding agents (Claude Code, Codex, OpenCode) in containerized environments.

    ## Authentication
    Currently no authentication required. API keys are managed via Credentials API.

    ## API Structure
    - **Public API** (`/api/v1/*`): Core functionality for external integrations
    - **Admin API** (`/api/v1/admin/*`): Platform management operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AgentBox API
  version: "1.0"
paths:
...
  /sessions:
    get:
      description: Get a list of all container sessions with optional filtering
      parameters:
      - description: Filter by status
        enum:
        - creating
        - running
        - stopped
        - error
        in: query
        name: status
        type: string
      - description: Filter by agent type
        in: query
        name: agent
        type: string
      - description: Number of results to return
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
                  type: array
              type: object
      summary: List sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Create a new container session for running an agent
      parameters:
      - description: Session configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Create a session
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: Stop and delete a session, removing its container
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Delete a session
      tags:
      - Sessions
    get:
      description: Get detailed information about a specific session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get a session
      tags:
      - Sessions
  /sessions/{id}/exec:
    post:
      consumes:
      - application/json
      description: Send a prompt to the agent running in the session and get the response
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Prompt to execute
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.ExecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.ExecResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Execute prompt in session
      tags:
      - Sessions
  /sessions/{id}/exec/stream:
    post:
      consumes:
      - application/json
      description: Send a prompt to the agent and receive real-time events via SSE
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Prompt to execute
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.ExecRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Execute prompt in session with SSE streaming (Codex only)
      tags:
      - Sessions
  /sessions/{id}/executions:
    get:
      description: Get the execution history for a session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Execution'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: List executions
      tags:
      - Sessions
  /sessions/{id}/executions/{execId}:
    get:
      description: Get details of a specific execution
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Execution ID
        in: path
        name: execId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Execution'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get an execution
      tags:
      - Sessions
  /sessions/{id}/logs:
    get:
      description: Get container logs for a session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  properties:
                    logs:
                      type: string
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get session logs
      tags:
      - Sessions
  /sessions/{id}/logs/stream:
    get:
      description: Get real-time container logs via Server-Sent Events
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Stream session logs (SSE)
      tags:
      - Sessions
  /sessions/{id}/reconnect:
    post:
      description: Attempt to reconnect to an existing container session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Reconnect to a session
      tags:
      - Sessions
  /sessions/{id}/start:
    post:
      description: Start a stopped session's container
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Start a session
      tags:
      - Sessions
  /sessions/{id}/stop:
    post:
      description: Stop a running session's container
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_session.Session'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Stop a session
      tags:
      - Sessions
...
swagger: "2.0"
...